query_llm_service/requirements.txt
fastapi==0.115.8
huggingface_hub==0.28.0
python-dotenv==1.0.1
Requests==2.32.3
transformers==4.48.1
